Working version, vectorized in z
Significant debugging was needed because due to a low initialization of the valence thermal energy (forgot the factor M_VB), the energy was becoming negative when a resonant electron below the fermi energy thermalized. This could theoretically still happen with extremely short pulses below the feri edge but should now throw an error at least in DEBUG mode.
The "check_bounds" function and "PhysicsError" were introduced for debugging.

Next steps can be radial integration and parallelization.